Remove unit test's dependence on external environment
