--- data/conf/hive-site.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-<configuration>
-    <property>
-        <name>hive.metastore.uris</name>
-        <value>thrift://192.168.90.150:9083</value>
-    </property>
-</configuration>
