+++ src/main/resources/log4j.properties
@@ -15,11 +15,9 @@
 #  See the License for the specific language governing permissions and
 # limitations under the License.
 ################################################################################
-log4j.rootLogger=DEBUG, console
+log4j.rootLogger=ERROR, console
 log4j.appender.console=org.apache.log4j.ConsoleAppender
 log4j.appender.console.layout=org.apache.log4j.PatternLayout
 log4j.appender.console.layout.ConversionPattern=%d %5p %t %-20c.%M:%L - %m%n
-log4j.logger.org.apache.flink=INFO
-log4j.logger.org.apache.hadoop=INFO
-log4j.logger.org.apache.calcite=INFO
-log4j.logger.org.apache.calcite.rel.metadata=DEBUG 
+log4j.logger.org.apache.calcite.rel.metadata=DEBUG
+log4j.logger.com.hw.flink.lineage=DEBUG
