--- src/main/scala/com/dtwave/flink/lineage/FlinkStreamProgramWithoutPhysical.scala
@@ -1,176 +1,0 @@
-package com.dtwave.flink.lineage
-
-
-import org.apache.calcite.plan.hep.HepMatchOrder
-import org.apache.flink.configuration.Configuration
-import org.apache.flink.table.api.config.OptimizerConfigOptions
-import org.apache.flink.table.planner.plan.nodes.FlinkConventions
-import org.apache.flink.table.planner.plan.optimize.program._
-import org.apache.flink.table.planner.plan.rules.FlinkStreamRuleSets
-
-/**
- * Defines a sequence of programs to optimize for stream table plan without physical optimize .
- *
- * @description: delete time_indicator, physical and physical_rewrite
- * @author: HamaWhite
- * @version: 1.0.0
- * @date: 2022/8/6 10:58 PM
- */
-
-object FlinkStreamProgramWithoutPhysical {
-
-  val SUBQUERY_REWRITE = "subquery_rewrite"
-  val TEMPORAL_JOIN_REWRITE = "temporal_join_rewrite"
-  val DECORRELATE = "decorrelate"
-  val DEFAULT_REWRITE = "default_rewrite"
-  val PREDICATE_PUSHDOWN = "predicate_pushdown"
-  val JOIN_REORDER = "join_reorder"
-  val PROJECT_REWRITE = "project_rewrite"
-  val LOGICAL = "logical"
-  val LOGICAL_REWRITE = "logical_rewrite"
-  val TIME_INDICATOR = "time_indicator"
-
-
-  def buildProgram(config: Configuration): FlinkChainedProgram[StreamOptimizeContext] = {
-    val chainedProgram = new FlinkChainedProgram[StreamOptimizeContext]()
-
-    // rewrite sub-queries to joins
-    chainedProgram.addLast(
-      SUBQUERY_REWRITE,
-      FlinkGroupProgramBuilder.newBuilder[StreamOptimizeContext]
-        // rewrite QueryOperationCatalogViewTable before rewriting sub-queries
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.TABLE_REF_RULES)
-          .build(), "convert table references before rewriting sub-queries to semi-join")
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.SEMI_JOIN_RULES)
-          .build(), "rewrite sub-queries to semi-join")
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_COLLECTION)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.TABLE_SUBQUERY_RULES)
-          .build(), "sub-queries remove")
-        // convert RelOptTableImpl (which exists in SubQuery before) to FlinkRelOptTable
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.TABLE_REF_RULES)
-          .build(), "convert table references after sub-queries removed")
-        .build())
-
-    // rewrite special temporal join plan
-    chainedProgram.addLast(
-      TEMPORAL_JOIN_REWRITE,
-      FlinkGroupProgramBuilder.newBuilder[StreamOptimizeContext]
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.EXPAND_PLAN_RULES)
-            .build(), "convert correlate to temporal table join")
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.POST_EXPAND_CLEAN_UP_RULES)
-            .build(), "convert enumerable table scan")
-        .build())
-
-    // query decorrelation
-    chainedProgram.addLast(DECORRELATE,
-      FlinkGroupProgramBuilder.newBuilder[StreamOptimizeContext]
-        // rewrite before decorrelation
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.PRE_DECORRELATION_RULES)
-            .build(), "pre-rewrite before decorrelation")
-        .addProgram(new FlinkDecorrelateProgram)
-        .build())
-
-    // default rewrite, includes: predicate simplification, expression reduction, window
-    // properties rewrite, etc.
-    chainedProgram.addLast(
-      DEFAULT_REWRITE,
-      FlinkHepRuleSetProgramBuilder.newBuilder
-        .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-        .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-        .add(FlinkStreamRuleSets.DEFAULT_REWRITE_RULES)
-        .build())
-
-    // rule based optimization: push down predicate(s) in where clause, so it only needs to read
-    // the required data
-    chainedProgram.addLast(
-      PREDICATE_PUSHDOWN,
-      FlinkGroupProgramBuilder.newBuilder[StreamOptimizeContext]
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_COLLECTION)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.FILTER_PREPARE_RULES)
-            .build(), "filter rules")
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.FILTER_TABLESCAN_PUSHDOWN_RULES)
-            .build(), "push predicate into table scan")
-        .addProgram(
-          FlinkHepRuleSetProgramBuilder.newBuilder
-            .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-            .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-            .add(FlinkStreamRuleSets.PRUNE_EMPTY_RULES)
-            .build(), "prune empty after predicate push down")
-        .build())
-
-    // join reorder
-    if (config.getBoolean(OptimizerConfigOptions.TABLE_OPTIMIZER_JOIN_REORDER_ENABLED)) chainedProgram.addLast(
-      JOIN_REORDER,
-      FlinkGroupProgramBuilder.newBuilder[StreamOptimizeContext]
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_COLLECTION)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.JOIN_REORDER_PREPARE_RULES)
-          .build(), "merge join into MultiJoin")
-        .addProgram(FlinkHepRuleSetProgramBuilder.newBuilder
-          .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-          .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-          .add(FlinkStreamRuleSets.JOIN_REORDER_RULES)
-          .build(), "do join reorder")
-        .build())
-
-    // project rewrite
-    chainedProgram.addLast(
-      PROJECT_REWRITE,
-      FlinkHepRuleSetProgramBuilder.newBuilder
-        .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_COLLECTION)
-        .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-        .add(FlinkStreamRuleSets.PROJECT_RULES)
-        .build())
-
-    // optimize the logical plan
-    chainedProgram.addLast(
-      LOGICAL,
-      FlinkVolcanoProgramBuilder.newBuilder
-        .add(FlinkStreamRuleSets.LOGICAL_OPT_RULES)
-        .setRequiredOutputTraits(Array(FlinkConventions.LOGICAL))
-        .build())
-
-    // logical rewrite
-    chainedProgram.addLast(
-      LOGICAL_REWRITE,
-      FlinkHepRuleSetProgramBuilder.newBuilder
-        .setHepRulesExecutionType(HEP_RULES_EXECUTION_TYPE.RULE_SEQUENCE)
-        .setHepMatchOrder(HepMatchOrder.BOTTOM_UP)
-        .add(FlinkStreamRuleSets.LOGICAL_REWRITE)
-        .build())
-
-    chainedProgram
-  }
-}
-
