+++ src/test/java/com/hw/flink/lineage/common/MySuffixFunction.java
@@ -1,0 +1,17 @@
+package com.hw.flink.lineage.common;
+
+import org.apache.flink.table.functions.ScalarFunction;
+
+/**
+ * @description: MySuffixFunction
+ * @author: HamaWhite
+ * @version: 1.0.0
+ * @date: 2022/8/24 3:35 PM
+ */
+public class MySuffixFunction extends ScalarFunction {
+
+    public String eval(String input) {
+        return input.concat("-HamaWhite");
+    }
+
+}
