+++ src/test/java/com/hw/flink/lineage/tablefuncion/MySplitFunction.java
@@ -1,0 +1,23 @@
+package com.hw.flink.lineage.tablefuncion;
+
+import org.apache.flink.table.annotation.DataTypeHint;
+import org.apache.flink.table.annotation.FunctionHint;
+import org.apache.flink.table.functions.TableFunction;
+import org.apache.flink.types.Row;
+
+/**
+ * @description: MySplitFunction
+ * @author: HamaWhite
+ * @version: 1.0.0
+ * @date: 2022/11/23 11:15 AM
+ */
+@FunctionHint(output = @DataTypeHint("ROW<word STRING, length INT>"))
+public class MySplitFunction extends TableFunction<Row> {
+
+    public void eval(String str) {
+        for (String s : str.split(" ")) {
+            // use collect(...) to emit a row
+            collect(Row.of(s, s.length()));
+        }
+    }
+}
